requirements.txt
openpyxl>=3.1.5
pandas>=3.0.1
python-dotenv>=1.0.0
requests>=2.32.0
uswid @ git+https://github.com/hughsie/python-uswid.git@ce25e04642655691698034cc0439e149b74d58d6
